In GUI.h, created variables for height & width of the buttons, and wxButton & wxTextCtrl for the array of buttons and the display. Added the method for onButtonClick event, no code currently to activate it. Also declared Event Table.
In GUI.cpp, added the space for event table, space for code for cosmetics (color, fonts, styles, etc.) and created new objects for the display and buttons. They work in that they exist in the program window, but the wxPoint & wxSize need to be adjusted to look correctly.

Add code to display the grid of buttons, code will meed to be adjusted to look correctly.