Attempted a few different implementations to get the onButtonClick method to work. Instead of doing each EVT_BUTTON in the event table, I used the EVT_COMMAND_RANGE to specify the ID range for my buttons.
Under the onButtonClick method, tried to do evt.getClientData or evt.getClientObject (and also getEventData and getEventObject) to display the button strings, but I couldn't figure out how to get the code to separate the individual evt commands. DisplayTextbox->AppendText(evt.GetClientData) and so forth didn't work.

I didn't want to use switch cases, because that's a lot of extra code, but that's what I have implemented here so far to at least get them to execute properly. I think after I complete the next lab (ButtonFactory), then I will have the buttons in a better spot to be able to actually 'go through' them or call them directly to append to the displayTextbox.

Have removed all previous code related to the arrayOfButtons that was commented out. Also created a region Button Creation to collapse that giant section of code.