Adjusted code of CalculatorProcessor.cpp to not have parameters for the functions, which allowed me to add void Execute() and put the methods in there.
Created a vector and a map for the wxButtons in the button factory as I am still trying to find the best way to connect the buttons and that data to everything. To CalculatorGUI I added a region ICommand that creates objects for each operand command and pushes them back into an IBaseCommand vector commands.